add extra logging
