+++ pkg/syncthing/monitor.go
@@ -2,12 +2,36 @@ package syncthing
 
 import (
 	"context"
+	"encoding/json"
 	"time"
+
+	"github.com/cloudnativedevelopment/cnd/pkg/log"
 )
 
-var consecutiveErrors = 0
+var consecutiveErrors = 1
 
 const maxConsecutiveErrors = 3
+
+func (s *Syncthing) isConnectedToRemote() bool {
+	body, err := s.GetFromAPI("rest/system/connections")
+	if err != nil {
+		log.Infof("error when getting connections from the api: %s", err)
+		return true
+	}
+
+	var conns syncthingConnections
+	if err := json.Unmarshal(body, &conns); err != nil {
+		log.Infof("error when unmarshalling response: %s", err)
+		return true
+	}
+
+	if val, ok := conns.Connections[s.RemoteDeviceID]; ok {
+		return val.Connected
+	}
+
+	log.Infof("RemoteDeviceID %s missing from the response", s.RemoteDeviceID)
+	return true
+}
 
 //Monitor verifies that syncthing is not in a disconnected state. If so, it sends a message to the
 // disconnected channel and exits
@@ -19,11 +43,14 @@ func (s *Syncthing) Monitor(ctx context.Context, disconnected chan struct{}) {
 			if !s.isConnectedToRemote() {
 				consecutiveErrors++
 				if consecutiveErrors > maxConsecutiveErrors {
+					log.Infof("not connected to syncthing, sending disconnect notification")
 					disconnected <- struct{}{}
 					return
 				}
+
+				log.Debugf("not connected to syncthing, try %d/%d", consecutiveErrors, maxConsecutiveErrors)
 			} else {
-				consecutiveErrors = 0
+				consecutiveErrors = 1
 			}
 
 		case <-ctx.Done():
